datacollector/requirements.txt
Django>=3.1.3
djangorestframework>=3.12.2
psycopg2>=2.8.6
djangorestframework-simplejwt>=4.5.0
django-cors-headers>=3.5.0
requests>=2.24.0
